% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{ISO2C Country Codes with names.}
\format{A character vector of 304 elements
\describe{
\item{countries}{Two Letter ISO2C Code for a country}
\item{names(countries)}{Full names of the country}
}}
\source{
http://databank.worldbank.org/data/reports.aspx?source=world-development-indicators
}
\usage{
countries
}
\description{
A named character vector of ISO2C 2 Character codes for country with the element name as the full name.
}
\examples{
  countries
}
\keyword{datasets}
