% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\docType{data}
\name{indicators}
\alias{indicators}
\title{WDI Indicator Codes with names.}
\format{A character vector of 17066 elements
\describe{
\item{indicators}{WDI Indicator Code}
\item{names(countries)}{Full names of the indicator}
}}
\source{
http://databank.worldbank.org/data/reports.aspx?source=world-development-indicators
}
\usage{
indicators
}
\description{
A named character vector of World Bank WDI Indicator codes for WDI Indicators with the element name as the full name.
}
\examples{
  indicators
}
\keyword{datasets}
